refactor: centralizar gerenciamento de estado
